+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.1]
+## [1.4.0]
 
 ### Added
 - Indicator in feed if the profile liked you or not.
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - "Interested in You" centered grid layout.
+- The horizontal scroll feature is now always enabled without any toggle.
 
 ### Fixed
 - Interested profiles fetch button fetching wrong profiles.